Deepen release page into workbench hook
